+++ tools/design_linters/framework/interfaces.py
@@ -19,199 +19,27 @@ Implementation: Enables plugin architecture with dynamic rule loading
 """
 
 import ast
-import re
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
 
 from .base_interfaces import BaseLintAnalyzer, BaseLintContext, BaseLintRule
+from .ignore_utils import has_file_level_ignore, should_ignore_node
 from .types import LintViolation, Severity
 
-
-def has_file_level_ignore(file_content: str, rule_id: str) -> bool:
-    """Check if file has file-level ignore directive for given rule."""
-    lines = file_content.split("\n")
-    for line in lines[:10]:  # Check only first 10 lines
-        if "# design-lint: ignore-file[" in line:
-            pattern = _extract_ignore_pattern(line, "ignore-file")
-            if pattern and _matches_rule_pattern(rule_id, pattern):
-                return True
-    return False
-
-
-def should_ignore_violation(violation: "LintViolation", file_content: str) -> bool:
-    """Check if a violation should be ignored based on inline directives."""
-    lines = file_content.split("\n")
-
-    # Check line-level ignore on same line
-    if violation.line <= len(lines):
-        line_content = lines[violation.line - 1]  # lines are 1-indexed
-        if "# design-lint: ignore[" in line_content:
-            pattern = _extract_ignore_pattern(line_content, "ignore")
-            if pattern and _matches_rule_pattern(violation.rule_id, pattern):
-                return True
-
-    # Check ignore-next-line directive on previous line
-    if violation.line > 1:
-        prev_line = lines[violation.line - 2]  # previous line
-        if "# design-lint: ignore-next-line" in prev_line:
-            return True
-
-    return False
-
-
-def _extract_ignore_pattern(line: str, directive_type: str) -> str | None:
-    """Extract ignore pattern from directive line."""
-    if directive_type == "ignore-file":
-        match = re.search(r"# design-lint: ignore-file\[([^\]]+)\]", line)
-    elif directive_type == "ignore":
-        match = re.search(r"# design-lint: ignore\[([^\]]+)\]", line)
-    else:
-        return None
-
-    return match.group(1) if match else None
-
-
-def _matches_rule_pattern(rule_id: str, pattern: str) -> bool:
-    """Check if rule ID matches ignore pattern."""
-    # Handle comma-separated patterns
-    patterns = [p.strip() for p in pattern.split(",")]
-
-    for p in patterns:
-        if p == rule_id:
-            return True
-
-        # Handle wildcard patterns like "literals.*"
-        if p.endswith(".*"):
-            prefix = p[:-2]
-            if rule_id.startswith(prefix + "."):
-                return True
-
-    return False
-
-
-def parse_ignore_directives(file_content: str, context: "LintContext") -> None:
-    """Parse ignore directives from file content and populate context."""
-    lines = file_content.split("\n")
-
-    for line_num, line in enumerate(lines, 1):
-        _process_file_level_ignore(line_num, line, context)
-        _process_line_level_ignore(line_num, line, context)
-        _process_ignore_next_line(line_num, line, context)
-
-
-def _process_file_level_ignore(line_num: int, line: str, context: "LintContext") -> None:
-    """Process file-level ignore directives."""
-    if line_num > 10 or "# design-lint: ignore-file[" not in line:
-        return
-    pattern = _extract_ignore_pattern(line, "ignore-file")
-    if pattern:
-        context.file_ignores.append(pattern)
-
-
-def _process_line_level_ignore(line_num: int, line: str, context: "LintContext") -> None:
-    """Process line-level ignore directives."""
-    if "# design-lint: ignore[" not in line:
-        return
-    pattern = _extract_ignore_pattern(line, "ignore")
-    if not pattern:
-        return
-    if line_num not in context.line_ignores:
-        context.line_ignores[line_num] = []
-    context.line_ignores[line_num].append(pattern)
-
-
-def _process_ignore_next_line(line_num: int, line: str, context: "LintContext") -> None:
-    """Process ignore-next-line directives."""
-    if "# design-lint: ignore-next-line" in line:
-        context.ignore_next_line.add(line_num + 1)  # Next line
-
-
-def should_ignore_node(node: ast.AST, file_content: str, rule_id: str) -> bool:
-    """Check if an AST node should be ignored for a specific rule."""
-    if not hasattr(node, "lineno"):
-        return False
-
-    line_num = node.lineno
-    lines = file_content.split("\n")
-
-    # Check if line is in ignore_next_line set (previous line had ignore-next-line)
-    if line_num in {
-        line_num
-        for line_num in range(1, len(lines) + 1)
-        if line_num > 1 and "# design-lint: ignore-next-line" in lines[line_num - 2]
-    }:
-        return True
-
-    # Check line-level ignore on same line
-    if line_num <= len(lines):
-        line_content = lines[line_num - 1]  # lines are 1-indexed
-        if "# design-lint: ignore[" in line_content:
-            pattern = _extract_ignore_pattern(line_content, "ignore")
-            if pattern and _matches_rule_pattern(rule_id, pattern):
-                return True
-
-    return False
+# All ignore utility functions are now imported from ignore_utils module
 
 
 class LintRule(BaseLintRule):
     """Abstract base class for all linting rules (Python-specific, inherits from BaseLintRule)."""
-
-    @property
-    @abstractmethod
-    def rule_id(self) -> str:
-        """Unique identifier for this rule."""
-        raise NotImplementedError("Subclasses must implement rule_id")
-
-    @property
-    @abstractmethod
-    def rule_name(self) -> str:
-        """Human-readable name for this rule."""
-        raise NotImplementedError("Subclasses must implement rule_name")
-
-    @property
-    @abstractmethod
-    def description(self) -> str:
-        """Description of what this rule checks."""
-        raise NotImplementedError("Subclasses must implement description")
-
-    @property
-    @abstractmethod
-    def severity(self) -> Severity:
-        """Default severity level for violations of this rule."""
-        raise NotImplementedError("Subclasses must implement severity")
-
-    @property
-    def categories(self) -> set[str]:
-        """Categories this rule belongs to (e.g., 'solid', 'style', 'complexity')."""
-        return set()
 
     @abstractmethod
     def check(self, context: "LintContext") -> list[LintViolation]:
         """Check for violations in the given context."""
         raise NotImplementedError("Subclasses must implement check")
 
-    def is_enabled(self, config: dict[str, Any] | None) -> bool:
-        """Check if this rule is enabled in the given configuration."""
-        if config is None:
-            return True
-        rules_config = config.get("rules", {})
-        rule_config = rules_config.get(self.rule_id, {})
-
-        # Check if there's a default rule enabled setting (used when specific rules are requested)
-        default_enabled = config.get("default_rule_enabled", True)
-        return bool(rule_config.get("enabled", default_enabled))
-
-    def get_configuration(self, config: dict[str, Any] | None) -> dict[str, Any]:
-        """Get configuration for this rule."""
-        if config is None:
-            return {}
-        rules_config = config.get("rules", {})
-        rule_config = rules_config.get(self.rule_id, {})
-        return dict(rule_config.get("config", {}))
-
-    def create_violation(
+    def create_violation_from_node(
         self,
         context: "LintContext",
         node: ast.AST,
@@ -221,17 +49,16 @@ class LintRule(BaseLintRule):
         suggestion: str | None = None,
         violation_context: dict[str, Any] | None = None,
     ) -> LintViolation:
-        """Helper method to create a violation with consistent structure."""
-        return LintViolation(
-            rule_id=self.rule_id,
-            file_path=str(context.file_path),
-            line=getattr(node, "lineno", 1),
-            column=getattr(node, "col_offset", 0),
-            severity=self.severity,
+        """Helper method to create a violation from an AST node."""
+        # Delegate to the base class method with extracted node info
+        return super().create_violation(
+            context,
             message=message,
             description=description,
+            line=getattr(node, "lineno", 1),
+            column=getattr(node, "col_offset", 0),
             suggestion=suggestion,
-            context=violation_context,
+            violation_context=violation_context,
         )
 
 
@@ -327,19 +154,9 @@ class FileBasedLintRule(LintRule):
 
     def check(self, context: "LintContext") -> list[LintViolation]:
         """Default implementation that checks entire file."""
-        violations: list[LintViolation] = []
+        from .ignore_utils import check_file_with_ignores
 
-        # Check for file-level ignore directives
-        if context.file_content and has_file_level_ignore(context.file_content, self.rule_id):
-            return violations
-
-        if context.file_path and context.file_content:
-            violations = self.check_file(context.file_path, context.file_content, context)
-            # Filter out violations on ignored lines
-            if context.file_content:
-                violations = [v for v in violations if not should_ignore_violation(v, context.file_content)]
-
-        return violations
+        return check_file_with_ignores(self, context, self.check_file)
 
 
 @dataclass
